Test.Cardano.Ledger.AllegraEraGen

Documentation

quantifyTL ∷ AllegraEraScript era ⇒ NativeScript era → Quantifier (NativeScript era)

unQuantifyTL ∷ AllegraEraScript era ⇒ Quantifier (NativeScript era) → NativeScript era

someLeaf ∷ ∀ era. (AllegraEraScript era, NativeScript era ~ Timelock era) ⇒ KeyHash 'Witness (EraCrypto era) → NativeScript era

Generate some Leaf Timelock (i.e. a Signature or TimeStart or TimeExpire).

Because we don't know how these "leaf scripts" will be situated in larger scripts (e.g. the script generated here might form part of a RequireAll or RequireMOf script) we must make sure that all timelocks generated here are valid for all slots.

To achieve this we arrange the timelock scripts like so: RequireAnyOf [ RequireAllOf [RequireTimeExpire k, RequireSignature x], RequireAllOf [RequireTimeStart k, RequireSignature x] ] where k is arbitrary. This means that regardless of slot, there will be a valid sub-branch of script.

genValidityInterval ∷ SlotNo → Gen ValidityInterval

Generates a trivial validity interval that is valid for the current slot.

Note: the validity interval must be a subset of all timelock script intervals that apply to the transaction. This depends on which generated scripts are actually required to validate the transaction (which is itself not always deterministic, e.g. 'RequireMOf n scripts').

A more sophisticated generator would compute which set of scripts would validate the transaction, and from that compute a minimal ValidityInterval that fits into all timelock slot ranges.

Orphan instances

Crypto c ⇒ EraGen (AllegraEra c)

Instance details

Methods

genGenesisValue ∷ GenEnv (AllegraEra c) → Gen (Value (AllegraEra c))

genEraTwoPhase3Arg ∷ [TwoPhase3ArgInfo (AllegraEra c)]

genEraTwoPhase2Arg ∷ [TwoPhase2ArgInfo (AllegraEra c)]

genEraTxBody ∷ GenEnv (AllegraEra c) → UTxO (AllegraEra c) → PParams (AllegraEra c) → SlotNo → Set (TxIn (EraCrypto (AllegraEra c))) → StrictSeq (TxOut (AllegraEra c)) → StrictSeq (TxCert (AllegraEra c)) → Withdrawals (EraCrypto (AllegraEra c)) → Coin → StrictMaybe (Update (AllegraEra c)) → StrictMaybe (AuxiliaryDataHash (EraCrypto (AllegraEra c))) → Gen (TxBody (AllegraEra c), [Script (AllegraEra c)])

genEraAuxiliaryData ∷ Constants → Gen (StrictMaybe (TxAuxData (AllegraEra c)))

updateEraTxBody ∷ UTxO (AllegraEra c) → PParams (AllegraEra c) → TxWits (AllegraEra c) → TxBody (AllegraEra c) → Coin → Set (TxIn (EraCrypto (AllegraEra c))) → TxOut (AllegraEra c) → TxBody (AllegraEra c)

addInputs ∷ TxBody (AllegraEra c) → Set (TxIn (EraCrypto (AllegraEra c))) → TxBody (AllegraEra c)

genEraPParamsUpdate ∷ Constants → PParams (AllegraEra c) → Gen (PParamsUpdate (AllegraEra c))

genEraPParams ∷ Constants → Gen (PParams (AllegraEra c))

genEraTxWits ∷ (UTxO (AllegraEra c), TxBody (AllegraEra c), ScriptInfo (AllegraEra c)) → Set (WitVKey 'Witness (EraCrypto (AllegraEra c))) → Map (ScriptHash (EraCrypto (AllegraEra c))) (Script (AllegraEra c)) → TxWits (AllegraEra c)

genEraGoodTxOut ∷ TxOut (AllegraEra c) → Bool

constructTx ∷ TxBody (AllegraEra c) → TxWits (AllegraEra c) → StrictMaybe (TxAuxData (AllegraEra c)) → Tx (AllegraEra c)

genEraScriptCost ∷ PParams (AllegraEra c) → Script (AllegraEra c) → Coin

genEraDone ∷ UTxO (AllegraEra c) → PParams (AllegraEra c) → Tx (AllegraEra c) → Gen (Tx (AllegraEra c))

genEraTweakBlock ∷ PParams (AllegraEra c) → Seq (Tx (AllegraEra c)) → Gen (Seq (Tx (AllegraEra c)))

hasFailedScripts ∷ Tx (AllegraEra c) → Bool

feeOrCollateral ∷ Tx (AllegraEra c) → UTxO (AllegraEra c) → Coin

Crypto c ⇒ MinGenTxout (AllegraEra c)

Instance details

Methods

calcEraMinUTxO ∷ TxOut (AllegraEra c) → PParams (AllegraEra c) → Coin

addValToTxOut ∷ Value (AllegraEra c) → TxOut (AllegraEra c) → TxOut (AllegraEra c)

genEraTxOut ∷ GenEnv (AllegraEra c) → Gen (Value (AllegraEra c)) → [Addr (EraCrypto (AllegraEra c))] → Gen [TxOut (AllegraEra c)]

Crypto c ⇒ ScriptClass (AllegraEra c)

Instance details

Methods

basescript ∷ Proxy (AllegraEra c) → KeyHash 'Witness (EraCrypto (AllegraEra c)) → Script (AllegraEra c)

isKey ∷ Proxy (AllegraEra c) → Script (AllegraEra c) → Maybe (KeyHash 'Witness (EraCrypto (AllegraEra c)))

isOnePhase ∷ Proxy (AllegraEra c) → Script (AllegraEra c) → Bool

quantify ∷ Proxy (AllegraEra c) → Script (AllegraEra c) → Quantifier (Script (AllegraEra c))

unQuantify ∷ Proxy (AllegraEra c) → Quantifier (Script (AllegraEra c)) → Script (AllegraEra c)