Test.Cardano.Ledger.Mary.Golden

Description

Documentation

goldenScaledMinDeposit ∷ TestTree

pid1 ∷ PolicyID StandardCrypto

pid2 ∷ PolicyID StandardCrypto

pid3 ∷ PolicyID StandardCrypto

smallName ∷ Word8 → AssetName

The small asset names have length one

smallestName ∷ AssetName

The smallest asset name has length zero

realName ∷ AssetName

try using a real asset name the way the CLI handles input

minUTxO ∷ Coin

This is the current value of the protocol parameter at the time this comment was written, namely one Ada.

largestName ∷ Word8 → AssetName

The largest asset names have length thirty-two