Test.Cardano.Ledger.MaryEraGen

Description

Export the EraGen instance for MaryEra, as well as some reusable functions for future Eras

Documentation

genMint ∷ Crypto c ⇒ Gen (MultiAsset c)

maryGenesisValue ∷ GenEnv era → Gen (MaryValue c)

Carefully crafted to apply in any Era where Value is MaryValue

policyIndex ∷ AllegraEraScript era ⇒ Map (PolicyID (EraCrypto era)) (NativeScript era)

Carefully crafted to apply in any Era where Value is MaryValue | This map allows us to lookup a minting policy by the policy ID.

addTokens ∷ ∀ era. (EraGen era, Value era ~ MaryValue (EraCrypto era)) ⇒ Proxy era → StrictSeq (TxOut era) → PParams era → MultiAsset (EraCrypto era) → StrictSeq (TxOut era) → Maybe (StrictSeq (TxOut era))

Carefully crafted to apply to any Era where Value is MaryValue We attempt to Add tokens to a non-empty list of transaction outputs. It will add them to the first output that has enough lovelace to meet the minUTxO requirment, if such an output exists.

Orphan instances

Crypto c ⇒ EraGen (MaryEra c)

Instance details

Methods

genGenesisValue ∷ GenEnv (MaryEra c) → Gen (Value (MaryEra c))

genEraTwoPhase3Arg ∷ [TwoPhase3ArgInfo (MaryEra c)]

genEraTwoPhase2Arg ∷ [TwoPhase2ArgInfo (MaryEra c)]

genEraTxBody ∷ GenEnv (MaryEra c) → UTxO (MaryEra c) → PParams (MaryEra c) → SlotNo → Set (TxIn (EraCrypto (MaryEra c))) → StrictSeq (TxOut (MaryEra c)) → StrictSeq (TxCert (MaryEra c)) → Withdrawals (EraCrypto (MaryEra c)) → Coin → StrictMaybe (Update (MaryEra c)) → StrictMaybe (AuxiliaryDataHash (EraCrypto (MaryEra c))) → Gen (TxBody (MaryEra c), [Script (MaryEra c)])

genEraAuxiliaryData ∷ Constants → Gen (StrictMaybe (TxAuxData (MaryEra c)))

updateEraTxBody ∷ UTxO (MaryEra c) → PParams (MaryEra c) → TxWits (MaryEra c) → TxBody (MaryEra c) → Coin → Set (TxIn (EraCrypto (MaryEra c))) → TxOut (MaryEra c) → TxBody (MaryEra c)

addInputs ∷ TxBody (MaryEra c) → Set (TxIn (EraCrypto (MaryEra c))) → TxBody (MaryEra c)

genEraPParamsUpdate ∷ Constants → PParams (MaryEra c) → Gen (PParamsUpdate (MaryEra c))

genEraPParams ∷ Constants → Gen (PParams (MaryEra c))

genEraTxWits ∷ (UTxO (MaryEra c), TxBody (MaryEra c), ScriptInfo (MaryEra c)) → Set (WitVKey 'Witness (EraCrypto (MaryEra c))) → Map (ScriptHash (EraCrypto (MaryEra c))) (Script (MaryEra c)) → TxWits (MaryEra c)

genEraGoodTxOut ∷ TxOut (MaryEra c) → Bool

constructTx ∷ TxBody (MaryEra c) → TxWits (MaryEra c) → StrictMaybe (TxAuxData (MaryEra c)) → Tx (MaryEra c)

genEraScriptCost ∷ PParams (MaryEra c) → Script (MaryEra c) → Coin

genEraDone ∷ UTxO (MaryEra c) → PParams (MaryEra c) → Tx (MaryEra c) → Gen (Tx (MaryEra c))

genEraTweakBlock ∷ PParams (MaryEra c) → Seq (Tx (MaryEra c)) → Gen (Seq (Tx (MaryEra c)))

hasFailedScripts ∷ Tx (MaryEra c) → Bool

feeOrCollateral ∷ Tx (MaryEra c) → UTxO (MaryEra c) → Coin

Crypto c ⇒ MinGenTxout (MaryEra c)

Instance details

Methods

calcEraMinUTxO ∷ TxOut (MaryEra c) → PParams (MaryEra c) → Coin

addValToTxOut ∷ Value (MaryEra c) → TxOut (MaryEra c) → TxOut (MaryEra c)

genEraTxOut ∷ GenEnv (MaryEra c) → Gen (Value (MaryEra c)) → [Addr (EraCrypto (MaryEra c))] → Gen [TxOut (MaryEra c)]

Crypto c ⇒ ScriptClass (MaryEra c)

Instance details

Methods

basescript ∷ Proxy (MaryEra c) → KeyHash 'Witness (EraCrypto (MaryEra c)) → Script (MaryEra c)

isKey ∷ Proxy (MaryEra c) → Script (MaryEra c) → Maybe (KeyHash 'Witness (EraCrypto (MaryEra c)))

isOnePhase ∷ Proxy (MaryEra c) → Script (MaryEra c) → Bool

quantify ∷ Proxy (MaryEra c) → Script (MaryEra c) → Quantifier (Script (MaryEra c))

unQuantify ∷ Proxy (MaryEra c) → Quantifier (Script (MaryEra c)) → Script (MaryEra c)

Split (MaryValue era)

Instance details

Methods

vsplit ∷ MaryValue era → Integer → ([MaryValue era], Coin)