Test.Cardano.Ledger.ShelleyMA.TestEra

Description

This module sets up a concrete Era for serialisation Tests. Users must make there own instances of the Era specific type families. The module provides some simple data types they can use, if their testing needs do not need complicated ones. Users make instances like these in the file that imports this file

data MyEra instance Era MyEra where type Crypto MyEra = TestCrypto type instance Value MyEra = ConcreteValue.Value MyEra type instance Script MyEra = TestScript type instance TxBody MyEra = Mary.TxBody MyEra

They may also skip the first three lines and Just use TestEra which is predefined.

Documentation

data TestEra

data TestCrypto

type TestValue = Value

data TestScript

data TestTxBody

type family Value era

A value is something which quantifies a transaction output.

type family Script era

Scripts which may lock transaction outputs in this era

type family TxBody era

The body of a transaction.