Test.Cardano.Ledger.EraBuffet

Documentation

data ShelleyEra

data MaryEra

data AllegraEra

type family Value era

A value is something which quantifies a transaction output.

type family Script era = (r ∷ Type) | r → era

Scripts which may lock transaction outputs in this era

data family TxBody era

The body of a transaction.

type family TxAuxData era = (r ∷ Type) | r → era

class (Typeable era, KnownNat (ProtVerLow era), KnownNat (ProtVerHigh era), ProtVerLow era <= ProtVerHigh era, MinVersion <= ProtVerLow era, MinVersion <= ProtVerHigh era, CmpNat (ProtVerLow era) MaxVersion ~ 'LT, CmpNat (ProtVerHigh era) MaxVersion ~ 'LT) ⇒ Era era